Move some classes into their own file
